Adding idea for bug
I'd like spotbugs to protect me from comparisons that are unsafe due to unboxing. We spent some time recently trying to debug a NullPointerException that was very hard to find. The problem was a bit like in potentiallyUnsafeButHardToFindCompareIntUsingUnboxing in the example code: the type of two sides of a comparison were "hidden" because they came from an instance of an external class, so a harmless int comparison was actually comparing an Integer that could be null against an int.
It would be great if spotbugs could warn against all unsafe comparisons that use unboxing.